Move HashSet and Dictionary out of Types.js
