+++ src/dotnet/Fable.Compiler/Transforms/Replacements.fs
@@ -2565,7 +2565,7 @@ let tryBaseConstructor com (ent: FSharpEntity) (memb: FSharpMemberOrFunctionOrVa
             | [Number _; IEqualityComparer], [_; eqComp] ->
                 [makeArray Any []; makeComparerFromEqualityComparer eqComp]
             | _ -> failwith "Unexpected dictionary constructor"
-        Some(makeCoreRef Any "Dictionary" "Types", args)
+        Some(makeCoreRef Any "Dictionary" "DictTypes", args)
     | Types.hashset ->
         let args =
             match FSharp2Fable.TypeHelpers.getArgTypes com memb, args with
@@ -2578,5 +2578,5 @@ let tryBaseConstructor com (ent: FSharpEntity) (memb: FSharpMemberOrFunctionOrVa
             | [IEqualityComparer], [eqComp] ->
                 [makeArray Any []; makeComparerFromEqualityComparer eqComp]
             | _ -> failwith "Unexpected hashset constructor"
-        Some(makeCoreRef Any "HashSet" "Types", args)
+        Some(makeCoreRef Any "HashSet" "DictTypes", args)
     | _ -> None
